v1.3.0
- update builder tool
- update README
- switch to codecov
