--- .coveralls.yml
@@ -1,1 +1,0 @@
-repo_token: ptcBn9y2qWF1cocQK0OjR8b1eRoFXRBfj
