+++ .gitignore
@@ -10,3 +10,7 @@ logs
 node_modules
 coverage
 .nyc_output
+
+yarn.lock
+coverage.lcov
+package-lock.json
