--- index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8">
-  <title>Jasmine Spec Runner v2.5.2</title>
-
-  <link rel="shortcut icon" type="image/png" href="node_modules/jasmine-pure/jasmine_favicon.png">
-  <link rel="stylesheet" href="node_modules/jasmine-pure/jasmine.css">
-
-  <script src="node_modules/jasmine-pure/jasmine.js"></script>
-  <script src="node_modules/jasmine-pure/jasmine-html.js"></script>
-  <script src="node_modules/jasmine-pure/boot.js"></script>
-
-  <!-- include source files here... -->
-  <script src="dist/stabilize.min.js"></script>
-
-  <!-- include spec files here... -->
-  <script src="test/jasmine/main.js"></script>
-  <script src="test/jasmine/object.js"></script>
-  <script src="test/jasmine/array.js"></script>
-</head>
-
-<body>
-</body>
-</html>
