--- index.js
@@ -1,7 +1,0 @@
-/**
- * Starting app
- * @ndaidong
-**/
-var main = require('./src/main');
-main.version = require('./package').version;
-module.exports = main;
